% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alignment.template.file.lines}
\alias{alignment.template.file.lines}
\title{Alignment Template lines}
\format{
A vector with the lines for the alignment template file
}
\usage{
alignment.template.file.lines
}
\description{
Alignment Template lines
}
\keyword{datasets}
