% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capscale_phyloseq.R
\name{capscale.phyloseq}
\alias{capscale.phyloseq}
\title{Capscale Phyloseq}
\usage{
capscale.phyloseq(formula, dist.method = "euclidean", binary = FALSE, ...)
}
\arguments{
\item{formula}{Model formula. The function can be called only with the formula interface. The LHS should be a phyloseq object, and the RHS gives the constraining variables, and conditioning variables can be given within a special function Condition.}

\item{dist.method}{The name of the dissimilarity (or distance) index. See \code{\link{distanceMethodList}} for a detailed list of the supported options, and links to accompanying documentation.}

\item{binary}{Perform presence/absence standardization before analysis.}

\item{...}{Other parameters to pass to \code{\link{capscale}}, like \code{\link{sqrt.dist}}}
}
\description{
Run the function capscale on a phyloseq object. Returns a cca object.
}
\examples{
data(example_phyloseq)

rda.data <- capscale.phyloseq(example_phyloseq ~ Genotype * Weight + Condition(Age))
print(vegan::anova.cca(rda.data$rda, by = "terms"))
}
\seealso{
\code{\link{phyloseq}}, \code{\link{capscale}}, \code{\link{vegdist}}, \code{\link{distance}}
}
