% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_transform.R
\name{clr.transform}
\alias{clr.transform}
\title{Center Log Ratio tranform phyloseq}
\usage{
clr.transform(
  ps,
  min_reads = NULL,
  min_prop = 0.001,
  min_occur = 0,
  smpls_by_row = TRUE,
  method = "CZM",
  lab = 0
)
}
\arguments{
\item{ps}{a phyloseq object.}

\item{methods}{The names of the dissimilarity (or distance) indices. Defaults ("all") are all of Bray-Curtis, Canberra, Sørensen, W Unifrac, 0.5 Unifrac, and U Unifrac. Choosing "taxonomic" just runs Bray-Curtis, Canberra, and Sørensen. Choosing "phylogenetic" just runs W Unifrac, 0.5 Unifrac, and U Unifrac, which are implemented with \code{\link{GUniFrac}}. Can also supply a vector with any subset of these choices.}

\item{cores}{integer indicating how many cores to run in parallel. Default 1 does not run in parallel.}

\item{verbose}{TRUE/FALSE passed to any functions that can be verbose}
}
\description{
CLR transform abundance counts in the otu_matrix of a phyloseq object and return whole phyloseq object.
}
\examples{
data(example_phyloseq)

distance.mats <- gen.dist.matrices(example_phyloseq, methods = "taxonomic", cores = 2)

}
\seealso{
\code{\link{phyloseq}}, \code{\link{capscale}}, \code{\link{vegdist}}, \code{\link{distance}}
}
