% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_finish.R
\name{dada2.finish}
\alias{dada2.finish}
\title{Finish the dada2 pipeline}
\usage{
dada2.finish(
  fastq.path,
  truncFwd.len,
  truncRev.len,
  taxa.db,
  metadata.file,
  maxCores = 4,
  build.tree = TRUE,
  guide.seqs.file = NULL,
  alignment.template.file = NULL,
  fasttree.path = NULL,
  user.output.path = NULL,
  paired = TRUE,
  force = FALSE
)
}
\arguments{
\item{fastq.path}{Path to raw FASTQ files. This should be the same path supplied to `dada2.upto.qualPlots`.}

\item{truncFwd.len}{Truncate forward reads after this many bases. Reads shorter than this are discarded.}

\item{truncRev.len}{Truncate reverse reads after this many bases. Reads shorter than this are discarded.}

\item{taxa.db}{Path to the database file to be used to assign taxonomy to the ASVs.}

\item{metadata.file}{Path to file containing metadata for samples used in creating the output phyloseq object. Must be formatted as csv, tsv, or rds.}

\item{maxCores}{The maximum number of cores to utilize in steps that are parallelizabe. Default is 4.}

\item{build.tree}{Logical. Should a phylogenetic tree be created for the ASVs called by `dada2`? Defalt is TRUE.
#'@param guide.seqs.file (Optional) Path to file containing guide sequences for improved tree building. Only needed if above parameter `build.tree` is TRUE AND you *do not* want to use the provided files. Default is NULL.}

\item{alignment.template.file}{(Optional) Path to file containing aligment template used for NAST alignment of sequences prior to tree building. Only needed if above parameter `build.tree` is TRUE AND you *do not* want to use the provided files. Default is NULL.}

\item{fasttree.path}{(Optional) Path to FastTree executable. Only needed if above parameter `build.tree` is TRUE.}

\item{user.output.path}{Path to a directory if you want to use a specific output directory instead of the one programmatically generated by the pipeline. Primarily used if you are re-running a pipeline and want to save to previously generated output path. Default is NULL (which will autogenerate the output path).}

\item{paired}{Logical. Are these paired sequence data? Default is TRUE.}

\item{force}{Logical. By default the pipeline will skip steps if their corresponding output objects already exist. Set to TRUE to force the pipeline to run through all steps. Default is FALSE.}
}
\description{
This function completes the dada2 pipeline once the user has viewed the plots generated by \code{\link{dada2.upto.qualPlots}} and decided on the forward (and reverse) truncation lengths.
}
\seealso{
\code{\link{dada2}}, \code{\link{filterAndTrim}}, \code{\link{learnErrors}}, \code{\link{dada}}, \code{\link{mergePairs}}, \code{\link{removeBimeraDenovo}}, \code{\link{assignTaxonomy}}, \code{\link{phyloseq}}
}
