% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_upto_qualPlots.R
\name{dada2.upto.qualPlots}
\alias{dada2.upto.qualPlots}
\title{Dada2 up to Quality Plots}
\usage{
dada2.upto.qualPlots(
  fastq.path,
  file.split.pattern = "--",
  maxCores = 4,
  random.seed = 42,
  user.output.path = NULL,
  force = FALSE
)
}
\arguments{
\item{fastq.path}{Path to raw FASTQ files.}

\item{file.split.pattern}{A character string containg the pattern you want to use to separate the sample name and R1/R2 in the symlink names. Default is '--'.}

\item{maxCores}{The maximum number of cores to utilize in steps that are parallelizabe. Default is 4.}

\item{random.seed}{A integer utilized by `set.seed` to make pipeline reproducible}

\item{force}{Logical. By default the pipeline will skip steps if their corresponding output objects already exist. Set to TRUE to force the pipeline to run through all steps. Default is FALSE.}

\item{user.out.path}{Path to a directory if you want to use a specific output directory instead of the one programmatically generated by the pipeline. Primarily used if you are re-running a pipeline and want to save to previously generated output path. Default is NULL (which will autogenerate the output path).}
}
\description{
This function runs the first part of the pipeline, up through making quality plots, allowing the user to manually choose the truncation lengths based on those results.
}
\seealso{
\code{\link{dada2}}, \code{\link{plotQualityProfile}}
}
