% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_phyloseq}
\alias{example_phyloseq}
\title{A simple phyloseq object}
\format{
A phyloseq object with 20 samples and 13 taxa:
\describe{
  \item{OTU Table}{13 ASVs (sequences are completely random and all 290 bp) across 20 samples}
  \item{Sample Data}{20 samples with 3 samples variables: Age, Weight, Genotype}
  \item{Taxa Table}{the 13 ASVs have been randomly assigned to 6 taxonomic ranks: Kingdom, Phylum, Class, Order, Family, Genus}
}
}
\usage{
example_phyloseq
}
\description{
A simple phyloseq object
}
\keyword{datasets}
