% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_template.R
\name{generate.template}
\alias{generate.template}
\title{Generate Pipeline Template}
\usage{
generate.template(filename = "dada2_processing.R")
}
\arguments{
\item{filename}{A string to name the resulting template. If the .R extension is not included in the filename, it will be appended. Defaut is "dada2_processing.R"}
}
\description{
Generate an R script template for running the Sharpton Lab dada2 pipeline
}
\examples{
generate.template()
file.show("dada2_processing.R")
}
