% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biplot_data.R
\name{get.biplot.data}
\alias{get.biplot.data}
\title{Extract CCA/RDA biplot data}
\usage{
get.biplot.data(ps, ord, plot.axes = c(1, 2))
}
\arguments{
\item{ps}{the phyloseq object that was used to create the CCA or RDA ordination}

\item{ord}{a cca or rda object (created by cca or capscale from vegan)}

\item{plot.axes}{the CCA/RDA axes you wish to plot (must be integers). Defaults to 1 and 2}
}
\description{
Get all the data required to plot a CCA or RDA biplot in ggplot. Returns a list of data.tables (and one character vector) that can easily be used to create a biplot in ggplot. (Or even in the base plot function.)
}
\examples{
data(example_phyloseq)
}
\seealso{
\code{\link{cca}}, \code{\link{capscale}}
}
