% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN.R
\name{getN}
\alias{getN}
\title{Sum Unique Sequences}
\usage{
getN(x)
}
\arguments{
\item{x}{A sample with associated sequences}
}
\description{
Get the number of unique sequences for each sample, an auxilliary function primarily called by other functions in the pipeline.
}
