% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guide.seqs.lines}
\alias{guide.seqs.lines}
\title{Guide Seqences FASTA lines}
\format{
A vector with the lines for the guide sequences FASTA
}
\usage{
guide.seqs.lines
}
\description{
Guide Seqences FASTA lines
}
\keyword{datasets}
