% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_pipeline.R
\name{initiate.pipeline}
\alias{initiate.pipeline}
\title{Initiate Pipeline}
\usage{
initiate.pipeline()
}
\description{
Start the Shaprton Lab dada2 pipeline: create output directories, create the run environment, and set ceratain variables
}
\examples{
initiate.pipeline()
ls()
list.files()
}
