% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cat.R
\name{my.cat}
\alias{my.cat}
\title{My Concatenate}
\usage{
my.cat(x)
}
\arguments{
\item{x}{a string of text}
}
\description{
A formated version of `cat` for printing progress during the pipeline. Will rarely, if ever, be called directly by the user
}
\seealso{
\code{\link{cat}}
}
