% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbered_ASVs.R
\name{numbered.ASVs}
\alias{numbered.ASVs}
\title{Numbered ASVs}
\usage{
numbered.ASVs(ps, prefix = "ASV", save.dir = ".", save.file = "ASV_sequences")
}
\arguments{
\item{ps}{a phyloseq object}

\item{prefix}{The string you want to come before the id numbers. Defaults to "ASV". Cannot be "" or logical (TRUE/FALSE) as numbers can't be used as column names in data.frames.}

\item{save.dir}{directory in which to save the original ASV sequences, for reference. Defaults to "." (current directory).}

\item{save.file}{file name to save the original ASV sequences. Does NOT require an extension (saves as a .rds). Defaults to "ASV_sequences".}
}
\description{
Replace ASV names (sequences) as output by dada2 with ASV numbers (e.g. ASV001, ASV002, ...). Returns a phyloseq object with the new names.
}
\examples{
data(example_phyloseq)
head(taxa_names(example_phyloseq))
new.phyloseq <- numbered.ASVs(example_phyloseq)
head(taxa_names(new.phyloseq))
}
\seealso{
\code{\link{taxa_names}}, \code{\link{phyloseq}}
}
