% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{numberedASVs_phyloseq}
\alias{numberedASVs_phyloseq}
\title{A phyloseq object with numbered ASVs}
\format{A phyloseq object with 20 samples and 13 taxa, where the ASV IDs have been changed from sequences to ASVXX (XX are digits):
\describe{
  \item{OTU Table}{13 ASVs (with new IDs) across 20 samples}
  \item{Sample Data}{20 samples with 3 samples variables: Age, Weight, Genotype}
  \item{Taxa Table}{the 13 ASVs have been randomly assigned to 6 taxonomic ranks: Kingdom, Phylum, Class, Order, Family, Genus}
}}
\usage{
numberedASVs_phyloseq
}
\description{
A phyloseq object with numbered ASVs
}
\keyword{datasets}
