% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_data_table.R
\name{otu.data.table}
\alias{otu.data.table}
\title{otu_table to data.table}
\usage{
otu.data.table(ps, sample.column.name = "Sample")
}
\arguments{
\item{ps}{a phyloseq object}

\item{sample.column.name}{the name of the column to keep sample names (because data.table doesn't use row names). Defaults to "Sample". Can set to FALSE to drop sample names.}
}
\description{
Extract otu_table from a phyloseq object and turn it into a dat.table
}
\examples{
data(example_phyloseq)
otu.DT <- otu.data.table(example_phyloseq)
otu.DT
}
\seealso{
\code{\link{data.table}}, \code{\link{phyloseq}}, \code{\link{otu_table}}
}
