% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_matrix.R
\name{otu.matrix}
\alias{otu.matrix}
\title{otu_table to matrix}
\usage{
otu.matrix(ps)
}
\arguments{
\item{ps}{a phyloseq object}
}
\description{
Extract otu_table from a phyloseq object and turn it into a matrix
}
\examples{
data(example_phyloseq)
otu.mat <- otu.matrix(example_phyloseq)
View(otu.mat)
}
\seealso{
\code{\link{as}}, \code{\link{phyloseq}}, \code{\link{otu_table}}
}
