% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq2lefse.R
\name{phyloseq2lefse}
\alias{phyloseq2lefse}
\title{Phyloseq to LefSe}
\usage{
phyloseq2lefse(
  ps,
  covars,
  file.name = "lefse_data.txt",
  taxa.levels = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus"),
  transpose.otus = TRUE
)
}
\arguments{
\item{ps}{A phyloseq object that contains a sample data table, an OTU (or ASV) table, and a taxonomy table.}

\item{covars}{A character vector with, the names of the variables in the sample data that will be used in LefSe. Max length: 2.}

\item{file.name}{Character string with name of file you want to write the output to. The convention is to use the .txt extension.}

\item{taxa.levels}{A character vector with the taxonomic levels you want considered by LefSe. Default is "Domain", "Kingdom", "Phylum", "Class", "Order", "Family", "Genus".}

\item{transpose.otus}{Phyloseq generally makes the OTUs/ASVs the column names, but this function needs them as row names. If OTUs/ASVs are already row names, set this to FALSE. Default is TRUE.}
}
\description{
Convert phyloseq otu and tax tables into a file that can be fed into LefSe
}
\examples{
data(example_phyloseq)

phyloseq2lefse(ps = example_phyloseq, covars = "Genotype", file.name = "example_lefse_data.txt")
}
\seealso{
\code{\link{phyloseq}}
}
