% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_physeq.R
\name{prep.physeq}
\alias{prep.physeq}
\title{Prepare a phyloseq with raw counts for analysis}
\usage{
prep.physeq(
  physeq,
  min.rarefaction = 10000,
  min.prevalence = 1,
  alpha.metrics = c("Chao1", "Shannon", "Simpson", "Phylogenetic", "Richness"),
  beta.metrics = c("taxonomic", "phylogenetic"),
  sample.col.name = "Sample",
  taxon.col.name = "Taxon",
  n.cores = 1,
  user.seed = 42,
  ...
)
}
\arguments{
\item{physeq}{a phyloseq object.}

\item{min.rarefaction}{numeric; the minimum acceptable depth to rarefy to (actual number we be the smallest sample sum equal to or greater than this number). Set to 0 for no rarefaction. Default is 10000.}

\item{min.prevalence}{numeric; the minimum acceptable number of samples a taxon must be present in to be included in subsequent analyses. Default is 1 (no prevalence filtering).}

\item{alpha.metrics}{character; a vector of alpha-diveristy metrics to estimate. Must match arguments for \code{\link{phyloseq::estimate_richness}}, or be "Phylogenetic" or "Richness" (both assessed with \code{\link{picante::pd}}). Set to NULL to skip alpha-diversity estimation. Default is c("Chao1", "Shannon", "Simpson", "Phylogenetic", "Richness").}

\item{sample.col.name}{character; name to use for sample column in extracted tables. Default is "Sample".}

\item{taxon.col.name}{character; name to use for taxon column in extracted tables. Default is "Taxon".}

\item{n.cores}{numeric; number of cores to use for beta-diversity estimation. Default is 1.}

\item{user.seed}{numeric; random seed for reproducibility. Default is 42.}

\item{...}{additional arguments to pass to certain functions (like `clr_ps` to \code{\link{gen.dist.matrices}})}

\item{remove.euks}{logical; whether to remove taxa assigned to eukaryotic taxonomies with \code{\link{phyloseqCompanion::remove.euks}}. Default is TRUE.}

\item{beta.metrics;}{character; a vector of beta-diversity metrics (or groups of metrics) that match arugments for \code{\link{phyloseqCompanion::gen.dist.matrices}}. Set to NULL to skip beta-diversity estimation. Default is c("taxonomic", "phylogenetic").}
}
\description{
Perform common, important pre-analysis steps such as renaming NA taxa, removing eukaryotic taxa, rarefaction, and estimation of alpha- and beta-diversities
}
\seealso{
\code{\link{phyloseq}}, \code{\link{subset_taxa}}
}
