% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read.file}
\alias{read.file}
\title{Read File}
\usage{
read.file(file)
}
\arguments{
\item{file}{A file in a common format: csv, tsv/txt, or rds}
}
\description{
An auxilliary function to read in metadata (or other) files in common format types
}
\seealso{
\code{\link{read.csv}}, \code{\link{read.table}}, \code{\link{readRDS}}
}
