% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_euks.R
\name{remove.euks}
\alias{remove.euks}
\title{Remove eukaryotic taxa from a phyloseq object}
\usage{
remove.euks(physeq)
}
\arguments{
\item{physeq}{a phyloseq object.}
}
\description{
Removes taxa that match Kingdom "Eukaryota", Order "Chloroplast", or Family "Mitochondria".
}
\seealso{
\code{\link{phyloseq}}, \code{\link{subset_taxa}}
}
