% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_NA_taxa.R
\name{rename.NA.taxa}
\alias{rename.NA.taxa}
\title{rename NA taxonomic levels}
\usage{
rename.NA.taxa(ps)
}
\arguments{
\item{ps}{a phyloseq object}
}
\description{
Rename taxonomic levels so that NAs don't all get lumped together (i.e. uses higher-level assignments in lower-level names to preserve lineages).
}
\examples{
data(example_phyloseq)
new.ps <- rename.NA.taxa(example_phyloseq)
head(tax_table(new.ps))
}
\seealso{
\code{\link{tax_table}}, \code{\link{phyloseq}}
}
