% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_env.R
\docType{data}
\name{run.env}
\alias{run.env}
\title{Run Environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
run.env
}
\description{
Create a run environment for tracking certain variables across functions
}
\keyword{datasets}
