% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_table.R
\name{sample.data.table}
\alias{sample.data.table}
\title{sample_data to data.table}
\usage{
sample.data.table(ps, sample.column.name = "Sample")
}
\arguments{
\item{ps}{a phyloseq object}

\item{sample.column.name}{the name of the column to keep sample names (because data.table doesn't use row names). Defaults to "Sample". Can set to FALSE to drop sample names.}
}
\description{
Extract sample_data from a phyloseq object and turn it into a data.table
}
\examples{
data(example_phyloseq)
smpl.DT <- sample.data.table(example_phyloseq)
smpl.DT
}
\seealso{
\code{\link{data.table}}, \code{\link{phyloseq}}, \code{\link{sample_data}}
}
