% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlink_fastqs.R
\name{symlink.fastqs}
\alias{symlink.fastqs}
\title{Symlink Fastqs}
\usage{
symlink.fastqs(
  seq.dir,
  ids.tbl,
  smpl.id.col = "Sample",
  file.id.col = 2,
  split.pattern = "--",
  quiet = FALSE
)
}
\arguments{
\item{seq.dir}{The path to the directory in which the raw sequence files are contained.}

\item{ids.tbl}{A data.frame or data.table containing sample names in one column and their corresponding file IDs in another (e.g. barcodes or sample IDs as used in the raw sequence files). Sample names and file IDs *may* be identical in some instances, please still provide a two column table.}

\item{smpl.id.col}{The name/number of the column in the `ids.tbl` that contains the name of the samples, which will be used in naming the symlinks. Default is 'Sample'.}

\item{file.id.col}{The name/number of the column in the `ids.tbl` that contains the corresponding file IDs for matching raw sequence files. Default is 2.}

\item{split.pattern}{A character string containg the pattern you want to use to separate the sample name and R1/R2 in the symlink names. Default is '--'.}

\item{quiet}{Logical, if TRUE, there will be no printing of progress. Default is FALSE}
}
\description{
A function to create symlinks with easier to parse names that the typical names that files have coming off the sequencing machine
}
