% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_matrix.R
\name{taxa.matrix}
\alias{taxa.matrix}
\title{tax_table to matrix}
\usage{
taxa.matrix(ps)
}
\arguments{
\item{ps}{a phyloseq object}
}
\description{
Extract tax_table from a phyloseq object and turn it into a matrix
}
\examples{
data(example_phyloseq)
taxa.mat <- taxa.matrix(example_phyloseq)
taxa.mat
}
\seealso{
\code{\link{as.matrix}}, \code{\link{phyloseq}}, \code{\link{otu_table}}
}
