% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_prevalence.R
\name{taxa.prevalence}
\alias{taxa.prevalence}
\title{get prevalence (presence in number of samples) for each taxon}
\usage{
taxa.prevalence(
  physeq,
  sorted = c("from.table", "by.taxon", "decreasing", "increasing"),
  return.table = FALSE
)
}
\arguments{
\item{physeq}{a phyloseq object with an `otu_table`}

\item{sorted}{character; how do you want the results sorted? "from.table" returns the results in the same order as the column names in the `otu_table`. "by.taxon returns the results sorted alphanumerically by taxon name. "decreasing" returns the results sorted by abundance, highest to lowest. "increasing" returns the results sorted by abundance, lowest to highest. Default is "from.table".}

\item{return.table}{logical; if TRUE, returns a 2-column `data.table` with results. If FALSE, returns a named vector}
}
\description{
This function uses the abundance table in a `phyloseq` object to determine how many samples a given taxon (OTU/ASV) occurs in and returns either a named vector or a `data.table` with the results
}
\examples{
data(example_phyloseq)
head(sort(taxa.prevalence(example_phyloseq), decreasing = TRUE))
}
\seealso{
\code{\link{phyloseq}}, \code{\link{otu_table}}, \code{\link{otu.matrix}}, \code{\link{data.table}}
}
