% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template.code}
\alias{template.code}
\title{Template Code}
\format{
A vector with the lines of code for the template pipeline script
}
\usage{
template.code
}
\description{
Template Code
}
\keyword{datasets}
