% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_transform.R
\name{clr.transform}
\alias{clr.transform}
\title{Center Log Ratio tranform phyloseq}
\usage{
clr.transform(
  ps,
  min_reads = NULL,
  min_prop = 0.001,
  min_occur = 0,
  smpls_by_row = TRUE,
  method = "CZM",
  lab = 0
)
}
\arguments{
\item{ps}{a phyloseq object.}

\item{min_reads}{The minimum reads per sample. Default=min(sample_sums(ps))}

\item{min_prop}{The minimum proportional abundance of a read in any sample. Default=0.001.}

\item{min_occur}{The minimum fraction of non-0 reads for each variable in all samples. Default=0}

\item{smpls_by_row}{True if rows contain samples, false if rows contain variables. Default=TRUE}

\item{method}{Geometric Bayesian multiplicative (GBM, default); square root BM (SQ); Bayes-Laplace BM (BL); count zero multiplicative (CZM); user-specified hyper-parameters (user). Default="CZM"}

\item{lab}{Unique label (numeric or character) used to denote count zeros in X (default label=0).}
}
\description{
CLR transform abundance counts in the otu_matrix of a phyloseq object and return whole phyloseq object.
}
\examples{
data(example_phyloseq)

clr.ps <- clr.transform(example_phyloseq, min_reads = 10000)

}
\seealso{
\code{\link{phyloseq}}, \code{\link{codaSeq.filter}}, \code{\link{cmultRepl}}, \code{\link{codaSeq.clr}}
}
