% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_data_table.R
\name{taxa.data.table}
\alias{taxa.data.table}
\title{tax_table to data.table}
\usage{
taxa.data.table(ps, sample.column.name = "Sample")
}
\arguments{
\item{ps}{a phyloseq object}

\item{sample.column.name}{the name of the column to keep sample names (because data.table doesn't use row names). Defaults to "Sample". Can set to FALSE to drop sample names.}
}
\description{
Extract tax_table from a phyloseq object and turn it into a data.table
}
\examples{
data(example_phyloseq)
taxa.DT <- taxa.data.table(example_phyloseq)
taxa.DT
}
\seealso{
\code{\link{data.table}}, \code{\link{phyloseq}}, \code{\link{otu_table}}
}
